+++ src/mock/offers.js
@@ -1,10 +1,12 @@
 import {Price} from './const.js';
 import {getRandomPositiveNumber} from './utils.js';
 
-const generateOffers = (type) => ({
+const generateMockOffers = (type) => ({
   id: crypto.randomUUID(),
   title: `Offer ${type}`,
   price: getRandomPositiveNumber(Price.MIN, Price.MAX / 10),
 });
 
-export {generateOffers};
+export {
+  generateMockOffers
+};
